actualizado crear autor

+++ ENTREGABLES/BACKEND/app/controllers/authors/delete-autor-by-id.js
@@ -1,7 +1,7 @@
 'use strict'
 
 const Joi = require('joi');
-const { deleteById } = require('../../repositories/author-repository')
+const { deleteById, findById } = require('../../repositories/author-repository')
 
 const schema = Joi.number().positive().required();
 
@@ -11,15 +11,14 @@ async function removeAuthorById(res, req) {
 
         await schema.validateAsync(idAuthor);
 
-        // const author = await deleteById(partseInt(idAuthor));
-        // console.log(idAuthor);
+        const author = await findById(idAuthor);
 
-        // if (author[0] === undefined) {
-        //     throw new Error("No se ha encontrado autor con ese id");
-        // }
+        if (author[0] === undefined) {
+            throw new Error("No se ha encontrado autor con ese id");
+        }
 
         await deleteById(parseInt(idAuthor));
-
+        console.log(author);
         res.status(200).send(`El autor ha sido eliminado correctamente.`);
     } catch(err) {
         res.status(400).send({ error: err.message });
